+++ cli/cli/commands/engine/status/status.go
@@ -17,7 +17,7 @@ var StatusCmd = &cobra.Command{
 func run(cmd *cobra.Command, args []string) error {
 	ctx := context.Background()
 
-	engineManager, err := engine_manager.NewEngineManager()
+	engineManager, err := engine_manager.NewEngineManager(ctx)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating an engine manager")
 	}