+++ cli/docs/changelog.md
@@ -18,6 +18,9 @@
 * Upgrade to engine 1.21.4 which understands Kubernetes cluster definitions
 * Execution Ids now serialize timestamp as `YYYY-MM-DDTHH-MM-SS-sss`
 
+### Fixes
+* Fixed a bug where pods would return before being started
+
 # 0.17.0
 ### Features
 * Added `test_cli_functionality` CircleCI test